Fix javascript-isms
